# api/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: ESIR2 Projet Simba - Smart Doodle
  version: 1.0.0
  title: Swagger Projet Simba
...
paths:
...
  /api/polls/{slug}/choices:
    post:
      tags:
        - Choices
      summary: Create new choices for a specific poll
      operationId: createChoices
      x-@id: evolvable:addChoiceToPoll
      parameters:
        - in: path
          name: slug
          schema:
            type: string
          required: true
        - in: query
          name: token
          schema:
            type: string
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChoiceCreationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Choices'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized


    delete:
      tags:
        - Choices
      operationId: deleteChoiceFromPoll
      x-@id: evolvable:deleteChoiceFromPoll
      summary: Delete a specific choice associated to a specific poll
      parameters:
        - in: path
          name: slug
          schema:
            type: string
          required: true
        - in: query
          name: token
          schema:
            type: string
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                choices:
                  type: array
                  items:
                    type: number
                    x-@type: evolvable:choiceId
                    format: int64
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '404':
          description: Not Found
  /api/polls/{slug}/choices/{choiceId}:
    put:
      tags:
        - Choices
      operationId: updateChoice
      x-@id: evolvable:updatePollChoice
      summary: Update a choice associated to a specific poll
      parameters:
        - in: path
          name: slug
          schema:
            type: string
          required: true
        - in: path
          name: choiceId
          schema:
            type: number
            format: int64
          required: true
        - in: query
          name: token
          schema:
            type: string
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChoiceUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Choice'
          links:
            delete:
              operationId: deleteChoiceFromPoll
              x-@relation: evolvable:rel/delete
              x-affiliation: parent
              parameters:
                slug: $request.path.slug
                choices:
                  - $response.body/#id
            update:
              operationId: updateChoice
              x-@relation: evolvable:rel/update
              x-affiliation: parent
              parameters:
                slug: $request.path.slug
                choiceId: $response.body/#id
        '401':
          description: 'Unauthorized'
        '404':
          description: Not Found
  /api/polls/{slug}/vote/{userId}:
    post:
      tags:
        - Votes
      operationId: vote
      parameters:
        - in: path
          name: slug
          schema:
            type: string
          required: true
        - in: path
          name: userId
          schema:
            type: number
            format: int64
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                choices:
                  type: array
                  items:
                    type: number
                    x-@type: ChoiceId
                    format: int64
      responses:
        '200':
          description: OK
          links:
            addMealPreference:
              operationId: vote
              x-@relation: evolvable:rel/nextVote
              parameters:
                slug: $request.path.slug
        '400':
          description: Bad Request
        '404':
          description: Not Found
...
components:
  schemas:
    User:
      type: object
      x-@type: evolvable:User
      required:
        - id
        - username
        - userChoices
        - userPolls
        - userMealPreferences
      properties:
        id:
          type: number
          format: long
        username:
          type: string
        userChoices:
          type: array
          default: []
          items:
            $ref: '#/components/schemas/Choice'
        userPolls:
          type: array
          default: []
          items:
            $ref: '#/components/schemas/Poll'
        userMealPreferences:
          type: array
          default: []
          items:
            $ref: '#/components/schemas/Choice'
...
    Choice:
      type: object
      x-@type: evolvable:Choice
      required:
        - id
        - startDate
        - endDate
        - users
      properties:
        id:
          x-@id:
            - evolvable:id
            - evolvable:choiceId
          x-@type:
            - evolvable:Id
            - evolvable:ChoiceId
          type: number
          format: long
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        users:
          $ref: '#/components/schemas/Users'
      links:
        delete:
          operationId: deleteChoiceFromPoll
          x-@relation: evolvable:rel/delete
          x-affiliation: parent
          parameters:
            slug: $request.path.slug
            choices:
              - $response.body/#id
        update:
          operationId: updateChoice
          x-@relation: evolvable:rel/update
          x-affiliation: parent
          parameters:
            slug: $request.path.slug
            choiceId: $response.body/#id
    ChoiceCreationRequest:
      type: array
      items:
        type: object
        required:
          - startDate
          - endDate
        properties:
          startDate:
            type: string
            format: date-time
          endDate:
            type: string
            format: date-time
          name:
            type: string
    ChoiceUpdateRequest:
      type: object
      required:
        - startDate
        - endDate
      properties:
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
    Choices:
      type: array
      x-@type: evolvable:Choices
      default: []
      items:
        $ref: '#/components/schemas/Choice'
    Comments:
      type: array
      x-@type: evolvable:Comments
      default: []
      items:
        $ref: '#/components/schemas/Comment'
    MealPreferences:
      type: array
      x-@type: evolvable:MealPreferences
      default: []
      items:
        $ref: '#/components/schemas/MealPreference'
    Users:
      type: array
      x-@type: evolvable:Users
      default: []
      items:
        $ref: '#/components/schemas/User'
    Poll:
      type: object
      x-@type: evolvable:Poll
      required:
        - id
        - title
        - location
        - has_meal
        - slug
        - createdAt
        - updatedAt
        - pollChoices
      properties:
        id:
          type: number
          format: long
        title:
          type: string
        location:
          type: string
        description:
          type: string
        has_meal:
          type: boolean
        slug:
          type: string
        slugAdmin:
          type: string
        tlkURL:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        pollChoices:
          $ref: '#/components/schemas/Choices'
        pollComments:
          $ref: '#/components/schemas/Comments'
        pollMealPreferences:
          $ref: '#/components/schemas/MealPreferences'
        padURL:
          type: string
...
    Comment:
      type: object
      x-@type: evolvable:Comment
      required:
        - id
        - content
        - user
      properties:
        id:
          type: number
          format: long
        content:
          type: string
        user:
          $ref: '#/components/schemas/User'
...
    MealPreference:
      type: object
      required:
        - id
        - content
        - user
      properties:
        id:
          type: number
          format: long
        content:
          type: string
        user:
          $ref: '#/components/schemas/User'


# json-LD context
...